backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.0
pdfplumber==0.10.3
Pillow==10.0.1
python-multipart==0.0.6
requests==2.31.0
google-api-core==2.11.0